backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.1
aiofiles>=23.0.0
# Database for session management
sqlalchemy>=2.0.0
# Optional: For production deployment
gunicorn>=21.0.0
# Task scheduling (optional for advanced features)
celery>=5.3.0
# Additional utilities
requests>=2.31.0

